Build-less sidebar js component
