+++ www/assets/js/index.js
@@ -1,2 +1,2 @@
-!function(t){var e=window.webpackJsonp;window.webpackJsonp=function(n,o,i){for(var a,u,c=0,s=[];c<n.length;c++)u=n[c],r[u]&&s.push(r[u][0]),r[u]=0;for(a in o)Object.prototype.hasOwnProperty.call(o,a)&&(t[a]=o[a]);for(e&&e(n,o,i);s.length;)s.shift()()};var n={},r={3:0};function o(e){if(n[e])return n[e].exports;var r=n[e]={i:e,l:!1,exports:{}};return t[e].call(r.exports,r,r.exports,o),r.l=!0,r.exports}o.e=function(t){var e=r[t];if(0===e)return new Promise(function(t){t()});if(e)return e[2];var n=new Promise(function(n,o){e=r[t]=[n,o]});e[2]=n;var i=document.getElementsByTagName("head")[0],a=document.createElement("script");a.type="text/javascript",a.charset="utf-8",a.async=!0,a.timeout=12e4,o.nc&&a.setAttribute("nonce",o.nc),a.src=o.p+"js/bundle-"+({0:"search",1:"code",2:"tooltip"}[t]||t)+".js?id="+{0:"b9dc7c6c802bf9882ff6",1:"0a6a4f1ea10fa159680e",2:"043f6d13541a8addd049"}[t];var u=setTimeout(c,12e4);function c(){a.onerror=a.onload=null,clearTimeout(u);var e=r[t];0!==e&&(e&&e[1](new Error("Loading chunk "+t+" failed.")),r[t]=void 0)}return a.onerror=a.onload=c,i.appendChild(a),n},o.m=t,o.c=n,o.d=function(t,e,n){o.o(t,e)||Object.defineProperty(t,e,{configurable:!1,enumerable:!0,get:n})},o.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return o.d(e,"a",e),e},o.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},o.p="/assets/",o.oe=function(t){throw console.error(t),t},o(o.s=0)}({"+Bt+":function(t,e){},0:function(t,e,n){n("vGYV"),n("SgXG"),n("P9yD"),n("+Bt+"),n("rhwI"),n("K4I0"),n("nFT7"),n("XgN2"),n("gpXz"),n("nHye"),n("m56c"),n("IFaM"),n("fDe5"),n("Md71"),n("Kn5K"),n("6J9z"),n("WgIl"),n("FnJk"),n("6314"),n("esmE"),n("2qlg"),n("w6Uh"),n("PprJ"),n("zlsc"),n("koDb"),n("i3Ly"),n("e0Aw"),n("vedj"),n("izEY"),t.exports=n("vg6f")},"2qlg":function(t,e){},6314:function(t,e){},"6J9z":function(t,e){},"7dJq":function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0});var r,o=(r=0,function(t){return new Promise(function(e){clearTimeout(r),r=setTimeout(e(),t)})});e.default=o},FnJk:function(t,e){},"Gnd+":function(t,e,n){"use strict";var r=function(){return function(t,e){if(Array.isArray(t))return t;if(Symbol.iterator in Object(t))return function(t,e){var n=[],r=!0,o=!1,i=void 0;try{for(var a,u=t[Symbol.iterator]();!(r=(a=u.next()).done)&&(n.push(a.value),!e||n.length!==e);r=!0);}catch(t){o=!0,i=t}finally{try{!r&&u.return&&u.return()}finally{if(o)throw i}}return n}(t,e);throw new TypeError("Invalid attempt to destructure non-iterable instance")}}(),o=a(n("sFUq")),i=a(n("gc1H"));function a(t){return t&&t.__esModule?t:{default:t}}(0,o.default)().then(function(){Promise.all([n.e(2).then(n.bind(null,"IDmg"))]).then(function(t){!function(t){var e=(0,i.default)(function(){!function(t){var e=document.querySelectorAll("[data-tooltip]");if(e.length)for(var n=0,r=e.length;n<r;n++){var o=e[n],i='<div class="tippy-inner | text text-small -background:black">'+o.getAttribute("data-tooltip")+"</div>",a={animation:"shift-away",arrow:!0,delay:[150,300],distance:10,duration:[200,150],interactive:!0,maxWidth:"25rem",theme:"kirby"};a.content=i,t(o,a)}}(t)},250);new MutationObserver(e).observe(document.documentElement,{childList:!0,subtree:!0,attributes:!1,characterData:!1}),e()}(r(t,1)[0].default)})})},IFaM:function(t,e){},Jucu:function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0}),e.on=function(t,e,n,r){t.addEventListener(e,function(e){for(var o=t.querySelectorAll(n),i=e.target,a=0,u=o.length;a<u;a++)for(var c=i,s=o[a];c&&c!==t;){if(c===s)return r.call(s,e);c=c.parentNode}})}},K4I0:function(t,e){},Kn5K:function(t,e){},Md71:function(t,e){},P9yD:function(t,e){},PSGk:function(t,e,n){"use strict";"kirbyConfig"in window&&window.kirbyConfig.assetsPath&&(n.p=window.kirbyConfig.assetsPath)},PprJ:function(t,e){},SgXG:function(t,e){},SldL:function(t,e){!function(e){"use strict";var n,r=Object.prototype,o=r.hasOwnProperty,i="function"==typeof Symbol?Symbol:{},a=i.iterator||"@@iterator",u=i.asyncIterator||"@@asyncIterator",c=i.toStringTag||"@@toStringTag",s="object"==typeof t,l=e.regeneratorRuntime;if(l)s&&(t.exports=l);else{(l=e.regeneratorRuntime=s?t.exports:{}).wrap=w;var d="suspendedStart",f="suspendedYield",h="executing",v="completed",m={},p={};p[a]=function(){return this};var y=Object.getPrototypeOf,g=y&&y(y(T([])));g&&g!==r&&o.call(g,a)&&(p=g);var b=k.prototype=E.prototype=Object.create(p);x.prototype=b.constructor=k,k.constructor=x,k[c]=x.displayName="GeneratorFunction",l.isGeneratorFunction=function(t){var e="function"==typeof t&&t.constructor;return!!e&&(e===x||"GeneratorFunction"===(e.displayName||e.name))},l.mark=function(t){return Object.setPrototypeOf?Object.setPrototypeOf(t,k):(t.__proto__=k,c in t||(t[c]="GeneratorFunction")),t.prototype=Object.create(b),t},l.awrap=function(t){return{__await:t}},_(P.prototype),P.prototype[u]=function(){return this},l.AsyncIterator=P,l.async=function(t,e,n,r){var o=new P(w(t,e,n,r));return l.isGeneratorFunction(e)?o:o.next().then(function(t){return t.done?t.value:o.next()})},_(b),b[c]="Generator",b[a]=function(){return this},b.toString=function(){return"[object Generator]"},l.keys=function(t){var e=[];for(var n in t)e.push(n);return e.reverse(),function n(){for(;e.length;){var r=e.pop();if(r in t)return n.value=r,n.done=!1,n}return n.done=!0,n}},l.values=T,A.prototype={constructor:A,reset:function(t){if(this.prev=0,this.next=0,this.sent=this._sent=n,this.done=!1,this.delegate=null,this.method="next",this.arg=n,this.tryEntries.forEach(j),!t)for(var e in this)"t"===e.charAt(0)&&o.call(this,e)&&!isNaN(+e.slice(1))&&(this[e]=n)},stop:function(){this.done=!0;var t=this.tryEntries[0].completion;if("throw"===t.type)throw t.arg;return this.rval},dispatchException:function(t){if(this.done)throw t;var e=this;function r(r,o){return u.type="throw",u.arg=t,e.next=r,o&&(e.method="next",e.arg=n),!!o}for(var i=this.tryEntries.length-1;i>=0;--i){var a=this.tryEntries[i],u=a.completion;if("root"===a.tryLoc)return r("end");if(a.tryLoc<=this.prev){var c=o.call(a,"catchLoc"),s=o.call(a,"finallyLoc");if(c&&s){if(this.prev<a.catchLoc)return r(a.catchLoc,!0);if(this.prev<a.finallyLoc)return r(a.finallyLoc)}else if(c){if(this.prev<a.catchLoc)return r(a.catchLoc,!0)}else{if(!s)throw new Error("try statement without catch or finally");if(this.prev<a.finallyLoc)return r(a.finallyLoc)}}}},abrupt:function(t,e){for(var n=this.tryEntries.length-1;n>=0;--n){var r=this.tryEntries[n];if(r.tryLoc<=this.prev&&o.call(r,"finallyLoc")&&this.prev<r.finallyLoc){var i=r;break}}i&&("break"===t||"continue"===t)&&i.tryLoc<=e&&e<=i.finallyLoc&&(i=null);var a=i?i.completion:{};return a.type=t,a.arg=e,i?(this.method="next",this.next=i.finallyLoc,m):this.complete(a)},complete:function(t,e){if("throw"===t.type)throw t.arg;return"break"===t.type||"continue"===t.type?this.next=t.arg:"return"===t.type?(this.rval=this.arg=t.arg,this.method="return",this.next="end"):"normal"===t.type&&e&&(this.next=e),m},finish:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var n=this.tryEntries[e];if(n.finallyLoc===t)return this.complete(n.completion,n.afterLoc),j(n),m}},catch:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var n=this.tryEntries[e];if(n.tryLoc===t){var r=n.completion;if("throw"===r.type){var o=r.arg;j(n)}return o}}throw new Error("illegal catch attempt")},delegateYield:function(t,e,r){return this.delegate={iterator:T(t),resultName:e,nextLoc:r},"next"===this.method&&(this.arg=n),m}}}function w(t,e,n,r){var o=e&&e.prototype instanceof E?e:E,i=Object.create(o.prototype),a=new A(r||[]);return i._invoke=function(t,e,n){var r=d;return function(o,i){if(r===h)throw new Error("Generator is already running");if(r===v){if("throw"===o)throw i;return C()}for(n.method=o,n.arg=i;;){var a=n.delegate;if(a){var u=O(a,n);if(u){if(u===m)continue;return u}}if("next"===n.method)n.sent=n._sent=n.arg;else if("throw"===n.method){if(r===d)throw r=v,n.arg;n.dispatchException(n.arg)}else"return"===n.method&&n.abrupt("return",n.arg);r=h;var c=L(t,e,n);if("normal"===c.type){if(r=n.done?v:f,c.arg===m)continue;return{value:c.arg,done:n.done}}"throw"===c.type&&(r=v,n.method="throw",n.arg=c.arg)}}}(t,n,a),i}function L(t,e,n){try{return{type:"normal",arg:t.call(e,n)}}catch(t){return{type:"throw",arg:t}}}function E(){}function x(){}function k(){}function _(t){["next","throw","return"].forEach(function(e){t[e]=function(t){return this._invoke(e,t)}})}function P(t){var e;this._invoke=function(n,r){function i(){return new Promise(function(e,i){!function e(n,r,i,a){var u=L(t[n],t,r);if("throw"!==u.type){var c=u.arg,s=c.value;return s&&"object"==typeof s&&o.call(s,"__await")?Promise.resolve(s.__await).then(function(t){e("next",t,i,a)},function(t){e("throw",t,i,a)}):Promise.resolve(s).then(function(t){c.value=t,i(c)},a)}a(u.arg)}(n,r,e,i)})}return e=e?e.then(i,i):i()}}function O(t,e){var r=t.iterator[e.method];if(r===n){if(e.delegate=null,"throw"===e.method){if(t.iterator.return&&(e.method="return",e.arg=n,O(t,e),"throw"===e.method))return m;e.method="throw",e.arg=new TypeError("The iterator does not provide a 'throw' method")}return m}var o=L(r,t.iterator,e.arg);if("throw"===o.type)return e.method="throw",e.arg=o.arg,e.delegate=null,m;var i=o.arg;return i?i.done?(e[t.resultName]=i.value,e.next=t.nextLoc,"return"!==e.method&&(e.method="next",e.arg=n),e.delegate=null,m):i:(e.method="throw",e.arg=new TypeError("iterator result is not an object"),e.delegate=null,m)}function S(t){var e={tryLoc:t[0]};1 in t&&(e.catchLoc=t[1]),2 in t&&(e.finallyLoc=t[2],e.afterLoc=t[3]),this.tryEntries.push(e)}function j(t){var e=t.completion||{};e.type="normal",delete e.arg,t.completion=e}function A(t){this.tryEntries=[{tryLoc:"root"}],t.forEach(S,this),this.reset(!0)}function T(t){if(t){var e=t[a];if(e)return e.call(t);if("function"==typeof t.next)return t;if(!isNaN(t.length)){var r=-1,i=function e(){for(;++r<t.length;)if(o.call(t,r))return e.value=t[r],e.done=!1,e;return e.value=n,e.done=!0,e};return i.next=i}}return{next:C}}function C(){return{value:n,done:!0}}}(function(){return this}()||Function("return this")())},WgIl:function(t,e){},XgN2:function(t,e){},Xxa5:function(t,e,n){t.exports=n("jyFz")},akH5:function(t,e,n){"use strict";var r=function(){return function(t,e){if(Array.isArray(t))return t;if(Symbol.iterator in Object(t))return function(t,e){var n=[],r=!0,o=!1,i=void 0;try{for(var a,u=t[Symbol.iterator]();!(r=(a=u.next()).done)&&(n.push(a.value),!e||n.length!==e);r=!0);}catch(t){o=!0,i=t}finally{try{!r&&u.return&&u.return()}finally{if(o)throw i}}return n}(t,e);throw new TypeError("Invalid attempt to destructure non-iterable instance")}}(),o=a(n("sFUq")),i=a(n("7dJq"));function a(t){return t&&t.__esModule?t:{default:t}}var u='<svg viewBox="0 0 5 11" width="5" height="11" aria-hidden="true"><path d="M2.96153846,5.41538462 L0,9.13846154 L1.35384615,10.1538462 L4.73846154,5.92307692 C4.82307692,5.75384615 4.90769231,5.58461538 4.90769231,5.41538462 C4.90769231,5.24615385 4.82307692,5.07692308 4.73846154,4.90769231 L1.35384615,0.676923077 L0,1.69230769 L2.96153846,5.41538462 Z"/></svg>';(0,o.default)(function(){var t=document.querySelector(".js-menu-search");if(t){var e=document.querySelector(".js-menu-search-input");Promise.all([n.e(0).then(n.bind(null,"l3sP")),n.e(0).then(n.bind(null,"S6+x"))]).then(function(n){var o=r(n,2),a=o[0],c=o[1],s=a("S7OGBIAJTV","d161a2f4cd2d69247c529a3371ad3050").initIndex("getkirby-3"),l="",d=[],f=new c(e,{list:[],filter:function(){return!0},sort:!1,item:function(t){var e=document.createElement("li");return"VIEW_ALL_RESULTS_UXfpeDFlmye9rXkXP5wd"===t.label?(e.classList.add("menu-search-view-all"),e.innerHTML="<strong>View all results</strong>"+u):"SEARCH_ERROR_UXfpeDFlmye9rXkXP5wd"===t.label?(e.classList.add("menu-search-error"),e.innerHTML="<strong>Sorry, an error occured. Please try advanced search instead.</strong>"+u):e.innerHTML="<strong>"+t.label+"</strong> <small>"+t.value+"</small>",e}});e.addEventListener("focus",function(){document.documentElement.classList.add("is-menu-search-open")}),e.addEventListener("blur",function(){document.documentElement.classList.remove("is-menu-search-open")}),e.addEventListener("keypress",function(e){(e.key&&"Enter"===e.key||13===e.keyCode)&&t.submit()}),e.addEventListener("keyup",function(){var t=e.value.trim();return t===l||(""===t?(l="",d=[],f.list=d,f.evaluate(),!0):(l=t,t.length<=2||void(0,i.default)(250).then(function(){var n={hitsPerPage:5},r=e.getAttribute("data-filters");r&&(n.filters=r),s.search(t,n,function(t,e){if(d=[],t)return"console"in window&&console.error("Quicksearch error",t),d.push({label:"SEARCH_ERROR_UXfpeDFlmye9rXkXP5wd",value:0}),f.list=d,f.evaluate(),void f.open();for(var n=0,r=e.hits.length;n<r;n++){var o=e.hits[n];d.push({label:o.title,value:o.objectID})}e.nbHits>0&&e.nbHits>e.hitsPerPage&&d.push({label:"VIEW_ALL_RESULTS_UXfpeDFlmye9rXkXP5wd",value:e.nbHits}),f.list=d,f.evaluate(),f.open()})})))}),e.addEventListener("awesomplete-select",function(e){e.preventDefault(),"VIEW_ALL_RESULTS_UXfpeDFlmye9rXkXP5wd"===e.text.label||"SEARCH_ERROR_UXfpeDFlmye9rXkXP5wd"===e.text.label?t.submit():window.location.href="/"+e.text.value})})}})},e0Aw:function(t,e){},eF2i:function(t,e,n){"use strict";var r,o=n("Jucu"),i=n("sFUq");function a(t){var e=t.querySelector("[aria-expanded]");if(!e)return!1;var n=t.querySelector(".js-sidebar-submenu");return!n.hidden?(n.hidden=!0,e.setAttribute("aria-expanded","false"),function(t){for(var e=0,n=t.querySelectorAll("a"),r=n.length;e<r;e++)n[e].setAttribute("tabindex","-1")}(n)):(e.setAttribute("aria-expanded","true"),n.hidden=!1,function(t){for(var e=0,n=t.querySelectorAll("a"),r=n.length;e<r;e++)n[e].setAttribute("tabindex","0")}(n)),!0}(0,((r=i)&&r.__esModule?r:{default:r}).default)().then(function(){var t=document.querySelector(".js-sidebar");t&&((0,o.on)(t,"keydown","a.sidebar-item-link",function(t){(t.key&&("ArrowRight"===t.key||"ArrowLeft"===t.key)||39===t.keyCode||37===t.keyCode)&&a(this.parentNode)&&t.preventDefault()}),(0,o.on)(t,"click",".js-sidebar-toggle",function(t){t.preventDefault(),a(this.parentNode)}))})},esmE:function(t,e){},fDe5:function(t,e){},gc1H:function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0}),e.default=function(t,e){var n=!1;return function(){n||(t(),n=!0,setTimeout(function(){n=!1},e))}}},gpXz:function(t,e){},i3Ly:function(t,e){},izEY:function(t,e){},jyFz:function(t,e,n){var r=function(){return this}()||Function("return this")(),o=r.regeneratorRuntime&&Object.getOwnPropertyNames(r).indexOf("regeneratorRuntime")>=0,i=o&&r.regeneratorRuntime;if(r.regeneratorRuntime=void 0,t.exports=n("SldL"),o)r.regeneratorRuntime=i;else try{delete r.regeneratorRuntime}catch(t){r.regeneratorRuntime=void 0}},koDb:function(t,e){},m56c:function(t,e){},nFT7:function(t,e){},nHye:function(t,e){},rhwI:function(t,e){},sFUq:function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0}),e.default=function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:function(){};return new Promise(function(e){"loading"!==document.readyState?(t(),e()):document.addEventListener("DOMContentLoaded",function(){t(),e()})})}},vGYV:function(t,e,n){"use strict";n("PSGk"),n("zlO1"),n("akH5"),n("eF2i"),n("Gnd+");var r,o=n("y8sO");(0,((r=o)&&r.__esModule?r:{default:r}).default)()},vedj:function(t,e){},vg6f:function(t,e){},w6Uh:function(t,e){},y8sO:function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0});var r,o,i=s(n("Xxa5")),a=function(){return function(t,e){if(Array.isArray(t))return t;if(Symbol.iterator in Object(t))return function(t,e){var n=[],r=!0,o=!1,i=void 0;try{for(var a,u=t[Symbol.iterator]();!(r=(a=u.next()).done)&&(n.push(a.value),!e||n.length!==e);r=!0);}catch(t){o=!0,i=t}finally{try{!r&&u.return&&u.return()}finally{if(o)throw i}}return n}(t,e);throw new TypeError("Invalid attempt to destructure non-iterable instance")}}(),u=(r=i.default.mark(function t(){var e,n,r,o;return i.default.wrap(function(t){for(;;)switch(t.prev=t.next){case 0:e=0,n=l.length;case 1:if(!(e<n)){t.next=12;break}if(r=l[e],"PRE"===(o=r.parentElement).tagName){t.next=6;break}return t.abrupt("continue",9);case 6:if(!o.className.match(/(^|\b)language-([^\b]|$)/)){t.next=8;break}return t.abrupt("continue",9);case 8:Prism.highlightElement(r);case 9:e++,t.next=1;break;case 12:case"end":return t.stop()}},t,this)}),o=function(){var t=r.apply(this,arguments);return new Promise(function(e,n){return function r(o,i){try{var a=t[o](i),u=a.value}catch(t){return void n(t)}if(!a.done)return Promise.resolve(u).then(function(t){r("next",t)},function(t){r("throw",t)});e(u)}("next")})},function(){return o.apply(this,arguments)}),c=s(n("gc1H"));function s(t){return t&&t.__esModule?t:{default:t}}var l=document.getElementsByTagName("code");function d(){var t=(0,c.default)(u,250);new MutationObserver(t).observe(document.documentElement,{childList:!0,subtree:!0,attributes:!1,characterData:!1}),t()}var f=function(t){var e=a(t,2)[1];Prism.languages.kirbytext=Prism.languages.extend("markdown",{}),Prism.languages.insertBefore("kirbytext","prolog",{kirbytag:{pattern:/\([a-z0-9_-]+:.*?\)/i,inside:{"kirbytag-bracket":/^\(|\)$/,"kirbytag-name":{pattern:/^[a-z0-9_-]+:/i},"kirbytag-attr":{pattern:/([^:]\s+)[a-z0-9_-]+:/i,lookbehind:!0}}}}),Prism.languages.kirbycontent={delimiter:/\n----\s*\n*/,property:{pattern:/(^\n*|\n----\s*\n*)[a-zA-Z0-9_\-\u0020]+:/,lookbehind:!0}},Prism.plugins.customClass.prefix("code-"),Prism.plugins.toolbar.registerButton("copy-to-clipboard",function(t){var n=document.createElement("a");n.classList.add("link-reset"),n.insertAdjacentHTML("beforeend",'<svg viewBox="0 0 16 16" width="16" height="16" class="icon"><path d="M10,4H2C1.4,4,1,4.4,1,5v10c0,0.6,0.4,1,1,1h8c0.6,0,1-0.4,1-1V5C11,4.4,10.6,4,10,4z"></path> <path data-color="color-2" d="M14,0H4v2h9v11h2V1C15,0.4,14.6,0,14,0z"></path></svg>');var r,o=document.createElement("span");function i(){setTimeout(function(){o.textContent="Copy"},5e3)}return o.textContent="Copy",n.appendChild(o),(r=new e(n,{text:function(){return t.code}})).on("success",function(){o.textContent="Copied!",i()}),r.on("error",function(){o.textContent="Press Ctrl+C/⌘+C to copy",i()}),n})};e.default=function(){for(var t,e=document.querySelectorAll('pre code[class^="language-"], pre code[class*=" language-"], pre.code > code'),r=/(?:^| )language(-[a-z])( |$)*/i,o=0,i=e.length;o<i&&(t=e[o]);o++)r.test(t.className)||t.classList.add("language-plaintext");Promise.all([n.e(1).then(n.bind(null,"h+jg")),n.e(1).then(n.bind(null,"TQvf"))]).then(f).then(d)}},zlO1:function(t,e,n){(function(){"use strict";"undefined"!=typeof document&&function(t){var e;function n(){e||(e=!0,t())}["interactive","complete"].indexOf(document.readyState)>=0?t():(e=!1,document.addEventListener("DOMContentLoaded",n,!1),window.addEventListener("load",n,!1))}(function(){var t=!0,e=!1,n=null,r={text:!0,search:!0,url:!0,tel:!0,email:!0,password:!0,number:!0,date:!0,month:!0,week:!0,time:!0,datetime:!0,"datetime-local":!0};function o(t){return!!(t&&t!==document&&"HTML"!==t.nodeName&&"BODY"!==t.nodeName&&"classList"in t&&"contains"in t.classList)}function i(t){t.classList.contains("focus-visible")||(t.classList.add("focus-visible"),t.setAttribute("data-focus-visible-added",""))}function a(e){t=!1}function u(){document.addEventListener("mousemove",c),document.addEventListener("mousedown",c),document.addEventListener("mouseup",c),document.addEventListener("pointermove",c),document.addEventListener("pointerdown",c),document.addEventListener("pointerup",c),document.addEventListener("touchmove",c),document.addEventListener("touchstart",c),document.addEventListener("touchend",c)}function c(e){"html"!==e.target.nodeName.toLowerCase()&&(t=!1,document.removeEventListener("mousemove",c),document.removeEventListener("mousedown",c),document.removeEventListener("mouseup",c),document.removeEventListener("pointermove",c),document.removeEventListener("pointerdown",c),document.removeEventListener("pointerup",c),document.removeEventListener("touchmove",c),document.removeEventListener("touchstart",c),document.removeEventListener("touchend",c))}document.addEventListener("keydown",function(e){o(document.activeElement)&&i(document.activeElement),t=!0},!0),document.addEventListener("mousedown",a,!0),document.addEventListener("pointerdown",a,!0),document.addEventListener("touchstart",a,!0),document.addEventListener("focus",function(e){var n,a,u;o(e.target)&&(t||(n=e.target,a=n.type,"INPUT"==(u=n.tagName)&&r[a]&&!n.readOnly||"TEXTAREA"==u&&!n.readOnly||n.isContentEditable))&&i(e.target)},!0),document.addEventListener("blur",function(t){var r;o(t.target)&&(t.target.classList.contains("focus-visible")||t.target.hasAttribute("data-focus-visible-added"))&&(e=!0,window.clearTimeout(n),n=window.setTimeout(function(){e=!1,window.clearTimeout(n)},100),(r=t.target).hasAttribute("data-focus-visible-added")&&(r.classList.remove("focus-visible"),r.removeAttribute("data-focus-visible-added")))},!0),document.addEventListener("visibilitychange",function(n){"hidden"==document.visibilityState&&(e&&(t=!0),u())},!0),u(),document.body.classList.add("js-focus-visible")})})()},zlsc:function(t,e){}});
+!function(t){var e=window.webpackJsonp;window.webpackJsonp=function(n,o,i){for(var a,u,c=0,s=[];c<n.length;c++)u=n[c],r[u]&&s.push(r[u][0]),r[u]=0;for(a in o)Object.prototype.hasOwnProperty.call(o,a)&&(t[a]=o[a]);for(e&&e(n,o,i);s.length;)s.shift()()};var n={},r={3:0};function o(e){if(n[e])return n[e].exports;var r=n[e]={i:e,l:!1,exports:{}};return t[e].call(r.exports,r,r.exports,o),r.l=!0,r.exports}o.e=function(t){var e=r[t];if(0===e)return new Promise(function(t){t()});if(e)return e[2];var n=new Promise(function(n,o){e=r[t]=[n,o]});e[2]=n;var i=document.getElementsByTagName("head")[0],a=document.createElement("script");a.type="text/javascript",a.charset="utf-8",a.async=!0,a.timeout=12e4,o.nc&&a.setAttribute("nonce",o.nc),a.src=o.p+"js/bundle-"+({0:"search",1:"code",2:"tooltip"}[t]||t)+".js?id="+{0:"b9dc7c6c802bf9882ff6",1:"0a6a4f1ea10fa159680e",2:"043f6d13541a8addd049"}[t];var u=setTimeout(c,12e4);function c(){a.onerror=a.onload=null,clearTimeout(u);var e=r[t];0!==e&&(e&&e[1](new Error("Loading chunk "+t+" failed.")),r[t]=void 0)}return a.onerror=a.onload=c,i.appendChild(a),n},o.m=t,o.c=n,o.d=function(t,e,n){o.o(t,e)||Object.defineProperty(t,e,{configurable:!1,enumerable:!0,get:n})},o.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return o.d(e,"a",e),e},o.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},o.p="/assets/",o.oe=function(t){throw console.error(t),t},o(o.s=0)}({"+Bt+":function(t,e){},0:function(t,e,n){n("vGYV"),n("SgXG"),n("P9yD"),n("+Bt+"),n("rhwI"),n("K4I0"),n("nFT7"),n("XgN2"),n("gpXz"),n("nHye"),n("m56c"),n("IFaM"),n("fDe5"),n("Md71"),n("Kn5K"),n("6J9z"),n("WgIl"),n("FnJk"),n("6314"),n("esmE"),n("2qlg"),n("w6Uh"),n("PprJ"),n("zlsc"),n("koDb"),n("i3Ly"),n("e0Aw"),n("vedj"),n("izEY"),t.exports=n("vg6f")},"2qlg":function(t,e){},6314:function(t,e){},"6J9z":function(t,e){},"7dJq":function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0});var r,o=(r=0,function(t){return new Promise(function(e){clearTimeout(r),r=setTimeout(e(),t)})});e.default=o},FnJk:function(t,e){},"Gnd+":function(t,e,n){"use strict";var r=function(){return function(t,e){if(Array.isArray(t))return t;if(Symbol.iterator in Object(t))return function(t,e){var n=[],r=!0,o=!1,i=void 0;try{for(var a,u=t[Symbol.iterator]();!(r=(a=u.next()).done)&&(n.push(a.value),!e||n.length!==e);r=!0);}catch(t){o=!0,i=t}finally{try{!r&&u.return&&u.return()}finally{if(o)throw i}}return n}(t,e);throw new TypeError("Invalid attempt to destructure non-iterable instance")}}(),o=a(n("sFUq")),i=a(n("gc1H"));function a(t){return t&&t.__esModule?t:{default:t}}(0,o.default)().then(function(){Promise.all([n.e(2).then(n.bind(null,"IDmg"))]).then(function(t){!function(t){var e=(0,i.default)(function(){!function(t){var e=document.querySelectorAll("[data-tooltip]");if(e.length)for(var n=0,r=e.length;n<r;n++){var o=e[n],i='<div class="tippy-inner | text text-small -background:black">'+o.getAttribute("data-tooltip")+"</div>",a={animation:"shift-away",arrow:!0,delay:[150,300],distance:10,duration:[200,150],interactive:!0,maxWidth:"25rem",theme:"kirby"};a.content=i,t(o,a)}}(t)},250);new MutationObserver(e).observe(document.documentElement,{childList:!0,subtree:!0,attributes:!1,characterData:!1}),e()}(r(t,1)[0].default)})})},IFaM:function(t,e){},K4I0:function(t,e){},Kn5K:function(t,e){},Md71:function(t,e){},P9yD:function(t,e){},PSGk:function(t,e,n){"use strict";"kirbyConfig"in window&&window.kirbyConfig.assetsPath&&(n.p=window.kirbyConfig.assetsPath)},PprJ:function(t,e){},SgXG:function(t,e){},SldL:function(t,e){!function(e){"use strict";var n,r=Object.prototype,o=r.hasOwnProperty,i="function"==typeof Symbol?Symbol:{},a=i.iterator||"@@iterator",u=i.asyncIterator||"@@asyncIterator",c=i.toStringTag||"@@toStringTag",s="object"==typeof t,l=e.regeneratorRuntime;if(l)s&&(t.exports=l);else{(l=e.regeneratorRuntime=s?t.exports:{}).wrap=w;var d="suspendedStart",f="suspendedYield",h="executing",v="completed",m={},p={};p[a]=function(){return this};var y=Object.getPrototypeOf,g=y&&y(y(A([])));g&&g!==r&&o.call(g,a)&&(p=g);var b=k.prototype=E.prototype=Object.create(p);x.prototype=b.constructor=k,k.constructor=x,k[c]=x.displayName="GeneratorFunction",l.isGeneratorFunction=function(t){var e="function"==typeof t&&t.constructor;return!!e&&(e===x||"GeneratorFunction"===(e.displayName||e.name))},l.mark=function(t){return Object.setPrototypeOf?Object.setPrototypeOf(t,k):(t.__proto__=k,c in t||(t[c]="GeneratorFunction")),t.prototype=Object.create(b),t},l.awrap=function(t){return{__await:t}},P(_.prototype),_.prototype[u]=function(){return this},l.AsyncIterator=_,l.async=function(t,e,n,r){var o=new _(w(t,e,n,r));return l.isGeneratorFunction(e)?o:o.next().then(function(t){return t.done?t.value:o.next()})},P(b),b[c]="Generator",b[a]=function(){return this},b.toString=function(){return"[object Generator]"},l.keys=function(t){var e=[];for(var n in t)e.push(n);return e.reverse(),function n(){for(;e.length;){var r=e.pop();if(r in t)return n.value=r,n.done=!1,n}return n.done=!0,n}},l.values=A,T.prototype={constructor:T,reset:function(t){if(this.prev=0,this.next=0,this.sent=this._sent=n,this.done=!1,this.delegate=null,this.method="next",this.arg=n,this.tryEntries.forEach(j),!t)for(var e in this)"t"===e.charAt(0)&&o.call(this,e)&&!isNaN(+e.slice(1))&&(this[e]=n)},stop:function(){this.done=!0;var t=this.tryEntries[0].completion;if("throw"===t.type)throw t.arg;return this.rval},dispatchException:function(t){if(this.done)throw t;var e=this;function r(r,o){return u.type="throw",u.arg=t,e.next=r,o&&(e.method="next",e.arg=n),!!o}for(var i=this.tryEntries.length-1;i>=0;--i){var a=this.tryEntries[i],u=a.completion;if("root"===a.tryLoc)return r("end");if(a.tryLoc<=this.prev){var c=o.call(a,"catchLoc"),s=o.call(a,"finallyLoc");if(c&&s){if(this.prev<a.catchLoc)return r(a.catchLoc,!0);if(this.prev<a.finallyLoc)return r(a.finallyLoc)}else if(c){if(this.prev<a.catchLoc)return r(a.catchLoc,!0)}else{if(!s)throw new Error("try statement without catch or finally");if(this.prev<a.finallyLoc)return r(a.finallyLoc)}}}},abrupt:function(t,e){for(var n=this.tryEntries.length-1;n>=0;--n){var r=this.tryEntries[n];if(r.tryLoc<=this.prev&&o.call(r,"finallyLoc")&&this.prev<r.finallyLoc){var i=r;break}}i&&("break"===t||"continue"===t)&&i.tryLoc<=e&&e<=i.finallyLoc&&(i=null);var a=i?i.completion:{};return a.type=t,a.arg=e,i?(this.method="next",this.next=i.finallyLoc,m):this.complete(a)},complete:function(t,e){if("throw"===t.type)throw t.arg;return"break"===t.type||"continue"===t.type?this.next=t.arg:"return"===t.type?(this.rval=this.arg=t.arg,this.method="return",this.next="end"):"normal"===t.type&&e&&(this.next=e),m},finish:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var n=this.tryEntries[e];if(n.finallyLoc===t)return this.complete(n.completion,n.afterLoc),j(n),m}},catch:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var n=this.tryEntries[e];if(n.tryLoc===t){var r=n.completion;if("throw"===r.type){var o=r.arg;j(n)}return o}}throw new Error("illegal catch attempt")},delegateYield:function(t,e,r){return this.delegate={iterator:A(t),resultName:e,nextLoc:r},"next"===this.method&&(this.arg=n),m}}}function w(t,e,n,r){var o=e&&e.prototype instanceof E?e:E,i=Object.create(o.prototype),a=new T(r||[]);return i._invoke=function(t,e,n){var r=d;return function(o,i){if(r===h)throw new Error("Generator is already running");if(r===v){if("throw"===o)throw i;return C()}for(n.method=o,n.arg=i;;){var a=n.delegate;if(a){var u=O(a,n);if(u){if(u===m)continue;return u}}if("next"===n.method)n.sent=n._sent=n.arg;else if("throw"===n.method){if(r===d)throw r=v,n.arg;n.dispatchException(n.arg)}else"return"===n.method&&n.abrupt("return",n.arg);r=h;var c=L(t,e,n);if("normal"===c.type){if(r=n.done?v:f,c.arg===m)continue;return{value:c.arg,done:n.done}}"throw"===c.type&&(r=v,n.method="throw",n.arg=c.arg)}}}(t,n,a),i}function L(t,e,n){try{return{type:"normal",arg:t.call(e,n)}}catch(t){return{type:"throw",arg:t}}}function E(){}function x(){}function k(){}function P(t){["next","throw","return"].forEach(function(e){t[e]=function(t){return this._invoke(e,t)}})}function _(t){var e;this._invoke=function(n,r){function i(){return new Promise(function(e,i){!function e(n,r,i,a){var u=L(t[n],t,r);if("throw"!==u.type){var c=u.arg,s=c.value;return s&&"object"==typeof s&&o.call(s,"__await")?Promise.resolve(s.__await).then(function(t){e("next",t,i,a)},function(t){e("throw",t,i,a)}):Promise.resolve(s).then(function(t){c.value=t,i(c)},a)}a(u.arg)}(n,r,e,i)})}return e=e?e.then(i,i):i()}}function O(t,e){var r=t.iterator[e.method];if(r===n){if(e.delegate=null,"throw"===e.method){if(t.iterator.return&&(e.method="return",e.arg=n,O(t,e),"throw"===e.method))return m;e.method="throw",e.arg=new TypeError("The iterator does not provide a 'throw' method")}return m}var o=L(r,t.iterator,e.arg);if("throw"===o.type)return e.method="throw",e.arg=o.arg,e.delegate=null,m;var i=o.arg;return i?i.done?(e[t.resultName]=i.value,e.next=t.nextLoc,"return"!==e.method&&(e.method="next",e.arg=n),e.delegate=null,m):i:(e.method="throw",e.arg=new TypeError("iterator result is not an object"),e.delegate=null,m)}function S(t){var e={tryLoc:t[0]};1 in t&&(e.catchLoc=t[1]),2 in t&&(e.finallyLoc=t[2],e.afterLoc=t[3]),this.tryEntries.push(e)}function j(t){var e=t.completion||{};e.type="normal",delete e.arg,t.completion=e}function T(t){this.tryEntries=[{tryLoc:"root"}],t.forEach(S,this),this.reset(!0)}function A(t){if(t){var e=t[a];if(e)return e.call(t);if("function"==typeof t.next)return t;if(!isNaN(t.length)){var r=-1,i=function e(){for(;++r<t.length;)if(o.call(t,r))return e.value=t[r],e.done=!1,e;return e.value=n,e.done=!0,e};return i.next=i}}return{next:C}}function C(){return{value:n,done:!0}}}(function(){return this}()||Function("return this")())},WgIl:function(t,e){},XgN2:function(t,e){},Xxa5:function(t,e,n){t.exports=n("jyFz")},akH5:function(t,e,n){"use strict";var r=function(){return function(t,e){if(Array.isArray(t))return t;if(Symbol.iterator in Object(t))return function(t,e){var n=[],r=!0,o=!1,i=void 0;try{for(var a,u=t[Symbol.iterator]();!(r=(a=u.next()).done)&&(n.push(a.value),!e||n.length!==e);r=!0);}catch(t){o=!0,i=t}finally{try{!r&&u.return&&u.return()}finally{if(o)throw i}}return n}(t,e);throw new TypeError("Invalid attempt to destructure non-iterable instance")}}(),o=a(n("sFUq")),i=a(n("7dJq"));function a(t){return t&&t.__esModule?t:{default:t}}var u='<svg viewBox="0 0 5 11" width="5" height="11" aria-hidden="true"><path d="M2.96153846,5.41538462 L0,9.13846154 L1.35384615,10.1538462 L4.73846154,5.92307692 C4.82307692,5.75384615 4.90769231,5.58461538 4.90769231,5.41538462 C4.90769231,5.24615385 4.82307692,5.07692308 4.73846154,4.90769231 L1.35384615,0.676923077 L0,1.69230769 L2.96153846,5.41538462 Z"/></svg>';(0,o.default)(function(){var t=document.querySelector(".js-menu-search");if(t){var e=document.querySelector(".js-menu-search-input");Promise.all([n.e(0).then(n.bind(null,"l3sP")),n.e(0).then(n.bind(null,"S6+x"))]).then(function(n){var o=r(n,2),a=o[0],c=o[1],s=a("S7OGBIAJTV","d161a2f4cd2d69247c529a3371ad3050").initIndex("getkirby-3"),l="",d=[],f=new c(e,{list:[],filter:function(){return!0},sort:!1,item:function(t){var e=document.createElement("li");return"VIEW_ALL_RESULTS_UXfpeDFlmye9rXkXP5wd"===t.label?(e.classList.add("menu-search-view-all"),e.innerHTML="<strong>View all results</strong>"+u):"SEARCH_ERROR_UXfpeDFlmye9rXkXP5wd"===t.label?(e.classList.add("menu-search-error"),e.innerHTML="<strong>Sorry, an error occured. Please try advanced search instead.</strong>"+u):e.innerHTML="<strong>"+t.label+"</strong> <small>"+t.value+"</small>",e}});e.addEventListener("focus",function(){document.documentElement.classList.add("is-menu-search-open")}),e.addEventListener("blur",function(){document.documentElement.classList.remove("is-menu-search-open")}),e.addEventListener("keypress",function(e){(e.key&&"Enter"===e.key||13===e.keyCode)&&t.submit()}),e.addEventListener("keyup",function(){var t=e.value.trim();return t===l||(""===t?(l="",d=[],f.list=d,f.evaluate(),!0):(l=t,t.length<=2||void(0,i.default)(250).then(function(){var n={hitsPerPage:5},r=e.getAttribute("data-filters");r&&(n.filters=r),s.search(t,n,function(t,e){if(d=[],t)return"console"in window&&console.error("Quicksearch error",t),d.push({label:"SEARCH_ERROR_UXfpeDFlmye9rXkXP5wd",value:0}),f.list=d,f.evaluate(),void f.open();for(var n=0,r=e.hits.length;n<r;n++){var o=e.hits[n];d.push({label:o.title,value:o.objectID})}e.nbHits>0&&e.nbHits>e.hitsPerPage&&d.push({label:"VIEW_ALL_RESULTS_UXfpeDFlmye9rXkXP5wd",value:e.nbHits}),f.list=d,f.evaluate(),f.open()})})))}),e.addEventListener("awesomplete-select",function(e){e.preventDefault(),"VIEW_ALL_RESULTS_UXfpeDFlmye9rXkXP5wd"===e.text.label||"SEARCH_ERROR_UXfpeDFlmye9rXkXP5wd"===e.text.label?t.submit():window.location.href="/"+e.text.value})})}})},e0Aw:function(t,e){},esmE:function(t,e){},fDe5:function(t,e){},gc1H:function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0}),e.default=function(t,e){var n=!1;return function(){n||(t(),n=!0,setTimeout(function(){n=!1},e))}}},gpXz:function(t,e){},i3Ly:function(t,e){},izEY:function(t,e){},jyFz:function(t,e,n){var r=function(){return this}()||Function("return this")(),o=r.regeneratorRuntime&&Object.getOwnPropertyNames(r).indexOf("regeneratorRuntime")>=0,i=o&&r.regeneratorRuntime;if(r.regeneratorRuntime=void 0,t.exports=n("SldL"),o)r.regeneratorRuntime=i;else try{delete r.regeneratorRuntime}catch(t){r.regeneratorRuntime=void 0}},koDb:function(t,e){},m56c:function(t,e){},nFT7:function(t,e){},nHye:function(t,e){},rhwI:function(t,e){},sFUq:function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0}),e.default=function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:function(){};return new Promise(function(e){"loading"!==document.readyState?(t(),e()):document.addEventListener("DOMContentLoaded",function(){t(),e()})})}},vGYV:function(t,e,n){"use strict";n("PSGk"),n("zlO1"),n("akH5"),n("Gnd+");var r,o=n("y8sO");(0,((r=o)&&r.__esModule?r:{default:r}).default)()},vedj:function(t,e){},vg6f:function(t,e){},w6Uh:function(t,e){},y8sO:function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0});var r,o,i=s(n("Xxa5")),a=function(){return function(t,e){if(Array.isArray(t))return t;if(Symbol.iterator in Object(t))return function(t,e){var n=[],r=!0,o=!1,i=void 0;try{for(var a,u=t[Symbol.iterator]();!(r=(a=u.next()).done)&&(n.push(a.value),!e||n.length!==e);r=!0);}catch(t){o=!0,i=t}finally{try{!r&&u.return&&u.return()}finally{if(o)throw i}}return n}(t,e);throw new TypeError("Invalid attempt to destructure non-iterable instance")}}(),u=(r=i.default.mark(function t(){var e,n,r,o;return i.default.wrap(function(t){for(;;)switch(t.prev=t.next){case 0:e=0,n=l.length;case 1:if(!(e<n)){t.next=12;break}if(r=l[e],"PRE"===(o=r.parentElement).tagName){t.next=6;break}return t.abrupt("continue",9);case 6:if(!o.className.match(/(^|\b)language-([^\b]|$)/)){t.next=8;break}return t.abrupt("continue",9);case 8:Prism.highlightElement(r);case 9:e++,t.next=1;break;case 12:case"end":return t.stop()}},t,this)}),o=function(){var t=r.apply(this,arguments);return new Promise(function(e,n){return function r(o,i){try{var a=t[o](i),u=a.value}catch(t){return void n(t)}if(!a.done)return Promise.resolve(u).then(function(t){r("next",t)},function(t){r("throw",t)});e(u)}("next")})},function(){return o.apply(this,arguments)}),c=s(n("gc1H"));function s(t){return t&&t.__esModule?t:{default:t}}var l=document.getElementsByTagName("code");function d(){var t=(0,c.default)(u,250);new MutationObserver(t).observe(document.documentElement,{childList:!0,subtree:!0,attributes:!1,characterData:!1}),t()}var f=function(t){var e=a(t,2)[1];Prism.languages.kirbytext=Prism.languages.extend("markdown",{}),Prism.languages.insertBefore("kirbytext","prolog",{kirbytag:{pattern:/\([a-z0-9_-]+:.*?\)/i,inside:{"kirbytag-bracket":/^\(|\)$/,"kirbytag-name":{pattern:/^[a-z0-9_-]+:/i},"kirbytag-attr":{pattern:/([^:]\s+)[a-z0-9_-]+:/i,lookbehind:!0}}}}),Prism.languages.kirbycontent={delimiter:/\n----\s*\n*/,property:{pattern:/(^\n*|\n----\s*\n*)[a-zA-Z0-9_\-\u0020]+:/,lookbehind:!0}},Prism.plugins.customClass.prefix("code-"),Prism.plugins.toolbar.registerButton("copy-to-clipboard",function(t){var n=document.createElement("a");n.classList.add("link-reset"),n.insertAdjacentHTML("beforeend",'<svg viewBox="0 0 16 16" width="16" height="16" class="icon"><path d="M10,4H2C1.4,4,1,4.4,1,5v10c0,0.6,0.4,1,1,1h8c0.6,0,1-0.4,1-1V5C11,4.4,10.6,4,10,4z"></path> <path data-color="color-2" d="M14,0H4v2h9v11h2V1C15,0.4,14.6,0,14,0z"></path></svg>');var r,o=document.createElement("span");function i(){setTimeout(function(){o.textContent="Copy"},5e3)}return o.textContent="Copy",n.appendChild(o),(r=new e(n,{text:function(){return t.code}})).on("success",function(){o.textContent="Copied!",i()}),r.on("error",function(){o.textContent="Press Ctrl+C/⌘+C to copy",i()}),n})};e.default=function(){for(var t,e=document.querySelectorAll('pre code[class^="language-"], pre code[class*=" language-"], pre.code > code'),r=/(?:^| )language(-[a-z])( |$)*/i,o=0,i=e.length;o<i&&(t=e[o]);o++)r.test(t.className)||t.classList.add("language-plaintext");Promise.all([n.e(1).then(n.bind(null,"h+jg")),n.e(1).then(n.bind(null,"TQvf"))]).then(f).then(d)}},zlO1:function(t,e,n){(function(){"use strict";"undefined"!=typeof document&&function(t){var e;function n(){e||(e=!0,t())}["interactive","complete"].indexOf(document.readyState)>=0?t():(e=!1,document.addEventListener("DOMContentLoaded",n,!1),window.addEventListener("load",n,!1))}(function(){var t=!0,e=!1,n=null,r={text:!0,search:!0,url:!0,tel:!0,email:!0,password:!0,number:!0,date:!0,month:!0,week:!0,time:!0,datetime:!0,"datetime-local":!0};function o(t){return!!(t&&t!==document&&"HTML"!==t.nodeName&&"BODY"!==t.nodeName&&"classList"in t&&"contains"in t.classList)}function i(t){t.classList.contains("focus-visible")||(t.classList.add("focus-visible"),t.setAttribute("data-focus-visible-added",""))}function a(e){t=!1}function u(){document.addEventListener("mousemove",c),document.addEventListener("mousedown",c),document.addEventListener("mouseup",c),document.addEventListener("pointermove",c),document.addEventListener("pointerdown",c),document.addEventListener("pointerup",c),document.addEventListener("touchmove",c),document.addEventListener("touchstart",c),document.addEventListener("touchend",c)}function c(e){"html"!==e.target.nodeName.toLowerCase()&&(t=!1,document.removeEventListener("mousemove",c),document.removeEventListener("mousedown",c),document.removeEventListener("mouseup",c),document.removeEventListener("pointermove",c),document.removeEventListener("pointerdown",c),document.removeEventListener("pointerup",c),document.removeEventListener("touchmove",c),document.removeEventListener("touchstart",c),document.removeEventListener("touchend",c))}document.addEventListener("keydown",function(e){o(document.activeElement)&&i(document.activeElement),t=!0},!0),document.addEventListener("mousedown",a,!0),document.addEventListener("pointerdown",a,!0),document.addEventListener("touchstart",a,!0),document.addEventListener("focus",function(e){var n,a,u;o(e.target)&&(t||(n=e.target,a=n.type,"INPUT"==(u=n.tagName)&&r[a]&&!n.readOnly||"TEXTAREA"==u&&!n.readOnly||n.isContentEditable))&&i(e.target)},!0),document.addEventListener("blur",function(t){var r;o(t.target)&&(t.target.classList.contains("focus-visible")||t.target.hasAttribute("data-focus-visible-added"))&&(e=!0,window.clearTimeout(n),n=window.setTimeout(function(){e=!1,window.clearTimeout(n)},100),(r=t.target).hasAttribute("data-focus-visible-added")&&(r.classList.remove("focus-visible"),r.removeAttribute("data-focus-visible-added")))},!0),document.addEventListener("visibilitychange",function(n){"hidden"==document.visibilityState&&(e&&(t=!0),u())},!0),u(),document.body.classList.add("js-focus-visible")})})()},zlsc:function(t,e){}});
 //# sourceMappingURL=index.js.map
