+++ www/assets/js/site.js
@@ -12,3 +12,11 @@ const menu = document.querySelector(".menu");
 if (menu) {
   new Menu(menu);
 }
+
+// Sidebar
+import Sidebar from "./components/sidebar.js"
+const sidebar = document.querySelector(".js-sidebar");
+
+if (sidebar) {
+  new Sidebar(sidebar);
+}
