+++ www/assets/mix-manifest.json
@@ -5,7 +5,7 @@
     "/js/bundle-code.js.map": "/js/bundle-code.js.map?id=1f4824c2adfb8e6e63d6",
     "/js/bundle-tooltip.js": "/js/bundle-tooltip.js?id=81a4eac4ad22105f6857",
     "/js/bundle-tooltip.js.map": "/js/bundle-tooltip.js.map?id=ef2f5267b93ae4ac4068",
-    "/js/index.js": "/js/index.js?id=63dbf368db037e132d59",
+    "/js/index.js": "/js/index.js?id=5efc54963af302155bf5",
     "/css/index.css": "/css/index.css?id=a43bed9e3a02d411f336",
     "/css/templates/archive.css": "/css/templates/archive.css?id=d3fe33b8452c5a10da9c",
     "/css/templates/buy.css": "/css/templates/buy.css?id=cac5746063868f0329fd",
@@ -35,7 +35,7 @@
     "/css/templates/text.css": "/css/templates/text.css?id=c265590bb71d15f06d1e",
     "/css/templates/try.css": "/css/templates/try.css?id=ddf925c7ca5e12567ed0",
     "/css/templates/why.css": "/css/templates/why.css?id=f4b2feed2d9fe592c103",
-    "/js/index.js.map": "/js/index.js.map?id=cf744a40c5d22a62422d",
+    "/js/index.js.map": "/js/index.js.map?id=506ae58716292ab626cf",
     "/css/index.css.map": "/css/index.css.map?id=abf0039b5ae3d5cf7e83",
     "/css/templates/archive.css.map": "/css/templates/archive.css.map?id=6c2694ea745675954f02",
     "/css/templates/buy.css.map": "/css/templates/buy.css.map?id=a1d87ca982c004494df7",
